+++ packages/web-components/components/selfie/src/selfie-capture/SelfieCapture.js
@@ -928,6 +928,7 @@ class SelfieCaptureScreen extends HTMLElement {
       'allow-agent-mode',
       'data-camera-error',
       'data-camera-ready',
+      'disable-image-tests',
       'hidden',
       'hide-back-to-host',
       'show-navigation',