+++ packages/web-components/components/smart-camera-web/src/SmartCameraWeb.js
@@ -61,8 +61,10 @@ class SmartCameraWeb extends HTMLElement {
 
   static get observedAttributes() {
     return [
+      'disable-image-tests',
       'document-capture-modes',
       'document-type',
+      'hide-back-of-id',
       'hide-back-to-host',
       'show-navigation',
       'hide-back-of-id',
@@ -72,6 +74,7 @@ class SmartCameraWeb extends HTMLElement {
 
   attributeChangedCallback(name) {
     switch (name) {
+      case 'disable-image-tests':
       case 'document-capture-modes':
       case 'document-type':
       case 'hide-back-of-id':